Heed compilation warnings. JIRA:CXX-3751
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@57389 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seq/unit_test/mapper_unit_test.cpp
@@ -711,7 +711,7 @@ void TestMapping_ThroughAlignments()
         "  Mapping through sparse-seg"
     };
 
-    for (int i = 0; i < sizeof(titles)/sizeof(titles[0]); i++) {
+    for (size_t i = 0; i < sizeof(titles)/sizeof(titles[0]); i++) {
         cout << titles[i] << endl;
         in >> MSerial_AsnText >> aln;
         auto_ptr<CSeq_loc_Mapper_Base> mapper(new CSeq_loc_Mapper_Base(aln, 0));
